% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  umc = c("utrecht", "maastricht", "amsterdam", "groningen", "rotterdam")
)
}
\arguments{
\item{umc}{University medical centre that provides the data. It can take the
the following values: "utrecht", "maastricht", "amsterdam", "groningen", "rotterdam"}
}
\value{
list with paths to the cleaned dataframes: visits and patients
}
\description{
Takes raw data from an sql output and format and clean it for further analysis This is the fist
function in this pipeline, it should be run first. It makes sure the variables are consistently
named in the same order and have the right format. It also check there are no missing or
nonsensical data. It ask the user to provide two raw data files: the one with GP contacts and
one with all patients. The cleaned data are automatically saved in the 'results' folder created
as a subfolder in the folder in which the raw data are loca ted.
}
