% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_contact.R
\docType{data}
\name{consult_contact}
\alias{consult_contact}
\title{consult_contact}
\format{
a character vector with 283 elements - contact types that are
considered a GP consultation
}
\usage{
consult_contact
}
\description{
List of eligible contact types
}
\details{
A list of contact types that are considered to be a GP consultation. It was
compiled by UMCU Utrecht. Contact Sander van Doorn for more details:
s.vandoorn at umcutrecht.nl
}
\keyword{datasets}
