% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominators.R
\name{denominators}
\alias{denominators}
\title{denominators}
\usage{
denominators(
  age_groups = c("0-19", "20-39", "40-59", "60-79", "80+"),
  exclude = c("0-19", "20-39")
)
}
\arguments{
\item{age_groups}{Definition of age groups stratification by age. Has to be
identical to 'age_groups' parameter used in the clean_data function !!}

\item{exclude}{Age groups to be excluded if we want to only consider older population}
}
\value{
the path to the resulting dataframes
}
\description{
Takes the cleaned patient data - saved by the clean_data function - to calculate
the population size for each period -  month, week and day - from 2016 to 2020.
The results are automatically saved in the results' folder created as a subfolder
in the folder in which the raw data are located.
}
\details{
There are no parameters
}
