% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_visits}
\alias{plot_visits}
\title{plot_visits}
\usage{
plot_visits(
  title = "This is a title",
  stratum = c("total", "age", "sex", "sex_age"),
  periodicity = c("m", "w", "d"),
  segment = "full",
  show_40plus = TRUE,
  path = NULL,
  point = 3,
  line = 1
)
}
\arguments{
\item{title}{Title of the plot}

\item{stratum}{How should the data be stratified, options are: 'sex': by sex, 'age': by age,
'sex_age': by age and sex, and "total": across all strata (the default).

Plot by sex will be shown on one panel, plots by age will be faceted across multiple panels}

\item{periodicity}{The period of counts, can either be 'd': day, 'w': week or 'm':
month (the default)}

\item{segment}{Time period to be plotted. By default, the data are plotted for the entire year,
with the pre-pandemic averages plotted against the 2020 data. However, when  plotting
daily data, it is recommended to use shorter period, which can be specified with
this argument. It takes a vector with two dates in the ymd format,
eg: c("2020-02-20", "2020-04-20").}

\item{show_40plus}{Specify whether you want to plot data for the entire population
or only for patients of 40 years and older}

\item{path}{Specify path to the  data to be plotted. Optional, if not provided
users will be prompted to select the data}

\item{point}{Size of the points on the resulting plot, defaults to 3}

\item{line}{Size of the points on the resulting plot, defaults to 1}
}
\description{
Plots the visit counts data (output of the n_visit function)
}
